refactor(slash): split slash.service from agent.service
agent.service was juggling four concerns: skill discovery, slash
expansion, always-active auto-injection, and the agent stream/persist
pipeline. Carve the two slash-shaped concerns out into a sibling
slash.service so the agent pipeline only knows about agent execution.
slashService is an internal dependency of agentService — composition
root injects it; AppEnv stays unchanged because no route consumes it
directly.
